requirements.txt
fastapi
uvicorn
requests
openai
beautifulsoup4
markdown
duckdb
gitpython
pillow
fuzzywuzzy
python-dotenv
